# swagger.yaml
openapi: 3.0.0

info:
  title: Node MySQL API
  version: 1.0.0
  description: API for managing accounts with JWT authentication and MySQL database.

servers:
  - url: http://localhost:4000
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        isVerified:
          type: boolean
    AuthenticateRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    AuthenticateResponse:
      allOf:
        - $ref: '#/components/schemas/Account'
        - type: object
          properties:
            jwtToken:
              type: string
            refreshToken:
              type: string
    RegisterRequest:
      type: object
      required:
        - title
        - firstName
        - lastName
        - email
        - password
        - confirmPassword
        - acceptTerms
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        acceptTerms:
          type: boolean
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
        - confirmPassword
      properties:
        token:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
    UpdateRequest:
      type: object
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        role:
          type: string

paths:
  /accounts/authenticate:
    post:
      tags:
        - Accounts
      summary: Authenticate user and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'

  /accounts/refresh-token:
    post:
      tags:
        - Accounts
      summary: Refresh JWT token
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'

  /accounts/revoke-token:
    post:
      tags:
        - Accounts
      summary: Revoke refresh token
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK

  /accounts/register:
    post:
      tags:
        - Accounts
      summary: Register new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK

  /accounts/verify-email:
    post:
      tags:
        - Accounts
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: OK

  /accounts/forgot-password:
    post:
      tags:
        - Accounts
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: OK

  /accounts/validate-reset-token:
    post:
      tags:
        - Accounts
      summary: Validate password reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: OK

  /accounts/reset-password:
    post:
      tags:
        - Accounts
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: OK

  /accounts:
    get:
      tags:
        - Accounts
      summary: Get all accounts (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      tags:
        - Accounts
      summary: Create new account (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK

  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    put:
      tags:
        - Accounts
      summary: Update account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags:
        - Accounts
      summary: Delete account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK